requirements.txt
google-adk==1.15.1
python-dotenv
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
Deprecated>=1.2.14
ffmpeg-python>=0.2.0
httpx>=0.27.0
